profile-api/requirements.txt
anyio==3.6.2
beanie==1.19.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
dnspython==2.3.0
fastapi==0.95.1
fastapi-jwt-auth==0.5.0
h11==0.14.0
idna==3.4
ifaddr==0.2.0
lazy-model==0.0.5
motor==3.1.2
py-eureka-client==0.11.7
pydantic==1.10.7
PyJWT==1.7.1
pymongo==4.3.3
python-decouple==3.8
requests==2.30.0
sniffio==1.3.0
starlette==0.26.1
toml==0.10.2
typing_extensions==4.5.0
urllib3==2.0.2
uuid6==2023.5.2
uvicorn==0.22.0
